==> explore/00-explore.ipynb <==
import pandas as pd

df_ = pd.read_excel(
    "../data/NIQU RMS - MTY.xlsx",
    sheet_name="7-BNA",
)

df_ = (
    df_.rename(
        columns={
            "Markets": "Mercado",
            "CATEGORIA": "Categoria",
            "M.PRESENTACION": "Presentacion",
            "M.SABOR": "Sabor",
            "Facts": "tipo",
            "Unnamed: 7": "valor",
        }
    )
    .drop(columns=["CANASTO"])
    .dropna()
)
df_

df = df_.copy()

month_map = {
    "Enero": "01",
    "Feb": "02",
    "Marzo": "03",
    "Abr": "04",
    "Mayo": "05",
    "Jun": "06",
    "Jul": "07",
    "Agosto": "08",
    "Set": "09",
    "Oct": "10",
    "Nov": "11",
    "Dic": "12",
}

month_series = df["Periods"].str.split(" ", expand=True)[0].map(month_map)
year_series = (
    df["Periods"].str.split(" ", expand=True)[1].apply(lambda year: f"20{year}")
)

date_series = year_series + "-" + month_series + "-01"

df["Fecha"] = pd.to_datetime(date_series)
df = df.drop(columns=["Periods"])
df

df.to_csv("../data/bebidas-no-alcoholicas.csv", index=False)

## Volumetría Simple

df["tipo"].value_counts()

df["Mercado"].value_counts()

df["Categoria"].value_counts()

df["Presentacion"].value_counts()

df["Sabor"].value_counts()

df.isna().sum()

## Pivot

df_pivot = (
    df.pivot_table(
        index=["Mercado", "Categoria", "Presentacion", "Fecha", "Sabor"],
        columns="tipo",
        values="valor",
        aggfunc="sum",
    )
    .reset_index()
    .dropna()
)

df_pivot["Mercado"] = (
    df_pivot["Mercado"]
    .str.strip()
    .str.removeprefix("Total_")
    .str.removesuffix("_MT _A")
    .str.strip()
)


df_pivot

df_pivot.isna().sum()

df_pivot.to_csv("../data/bebidas-no-alcoholicas-pivot.csv", index=False)
