==> explore/01-insights.ipynb <==
import pandas as pd
import plotly.graph_objects as go

df_ = pd.read_csv("../data/bebidas-no-alcoholicas-pivot.csv")
df_["Fecha"] = pd.to_datetime(df_["Fecha"])
df_

df_["Categoria"].value_counts()

df_["Mercado"].value_counts()

## Macroeconomico

### Stats

df = df_.copy().query("Mercado == 'Total_Mexico_MT _A'")

# 2022
df_2022 = df.query("Fecha.dt.year == 2022")
print(f"2022: ${df_2022['Vtas Valor'].sum():,.2f}")

# 2023
df_2023 = df.query("Fecha.dt.year == 2023")
print(f"2023: ${df_2023['Vtas Valor'].sum():,.2f}")

# 2024
df_2024 = df.query("Fecha.dt.year == 2024")
print(f"2024: ${df_2024['Vtas Valor'].sum():,.2f}")


df = df_.copy().query("Mercado == 'Total_Mexico_MT _A'")

# 2022
df_2022 = df.query("Fecha.dt.year == 2022")
print(f"2022: {df_2022['Vtas Unds'].sum():,.0f}")

# 2023
df_2023 = df.query("Fecha.dt.year == 2023")
print(f"2023: {df_2023['Vtas Unds'].sum():,.0f}")

# 2024
df_2024 = df.query("Fecha.dt.year == 2024")
print(f"2024: {df_2024['Vtas Unds'].sum():,.0f}")


df_["Sabor"].value_counts()

df = df_.copy().query("Mercado == 'Total_Mexico_MT _A'").query("Sabor == 'COLA'")

# 2022
df_2022 = df.query("Fecha.dt.year == 2022")
print(f"2022: {df_2022['Vtas Valor'].sum():,.2f}")

# 2023
df_2023 = df.query("Fecha.dt.year == 2023")
print(f"2023: {df_2023['Vtas Valor'].sum():,.2f}")

# 2024
df_2024 = df.query("Fecha.dt.year == 2024")
print(f"2024: {df_2024['Vtas Valor'].sum():,.2f}")


234_043_978_905.74 / 460_999_841_096.41

df_["Categoria"].value_counts()

df_.query("Presentacion == 'NOT APPLICABLE'").Categoria.unique()

### By month

MONTH_NAMES = {
    1: "Enero",
    2: "Febrero",
    3: "Marzo",
    4: "Abril",
    5: "Mayo",
    6: "Junio",
    7: "Julio",
    8: "Agosto",
    9: "Septiembre",
    10: "Octubre",
    11: "Noviembre",
    12: "Diciembre",
}


df = df_.copy().query("Mercado == 'Total_Mexico_MT _A'").query("Fecha.dt.year == 2024")
df["month"] = pd.to_datetime(df["Fecha"]).dt.month
df = df.groupby(["month"])["Vtas Valor"].sum().reset_index()
df["month"] = df["month"].map(MONTH_NAMES)
df

x = df["month"]
y = df["Vtas Valor"]

MONTH_COLORS = {
    1: "#7aa1f4",
    2: "#7aa1f4",
    3: "#7aa1f4",
    4: "#7aa1f4",
    5: "#ffd771",
    6: "#ffd771",
    7: "#7aa1f4",
    8: "#7aa1f4",
    9: "#7aa1f4",
    10: "#7aa1f4",
    11: "#7aa1f4",
    12: "#7aa1f4",
}


text = (df["Vtas Valor"] / 1_000_000_000).apply(lambda x: f"${x:,.2f}B")
# Create bar chart
fig = go.Figure(
    go.Bar(
        x=x,
        y=y,
        marker=dict(color=list(MONTH_COLORS.values())),
        text=text,
    )
)

fig.update_traces(textposition="outside")

title = "Ventas de Bebidas No Alcohólicas\n<br><i>por Mes en 2024</i>"
fig.update_layout(
    title=title,
    title_x=0.5,
    yaxis_title="Ventas ($)",
    barcornerradius=5,
    paper_bgcolor="rgba(0,0,0,0)",  # Transparent background
    plot_bgcolor="rgba(0,0,0,0)",
)

fig.update_xaxes(tickangle=-45)
fig.write_image("../img/00-macro/00-ventas-mes.png", scale=3)

fig.show()

### By presentation

df = df_.copy()
df = (
    df.groupby(["Presentacion"])["Vtas Valor"]
    .sum()
    .reset_index()
    .sort_values(by="Vtas Valor", ascending=False)
    .reset_index(drop=True)
)

df["percentage"] = df["Vtas Valor"] / df["Vtas Valor"].sum() * 100
df["percentage"] = df["percentage"].round(2)

df

labels = df["Presentacion"]
values = df["percentage"]

PRES_COLORS = {
    1: "#ffd771",
    2: "gray",
    3: "#7aa1f4",
    4: "#7aa1f4",
    5: "#7aa1f4",
    6: "#7aa1f4",
    7: "#7aa1f4",
}


fig = go.Figure(
    data=[
        go.Pie(
            labels=labels,
            values=values,
            pull=[0.2, 0, 0, 0, 0, 0, 0],
            hole=0.3,
            marker=dict(colors=list(PRES_COLORS.values())),
        )
    ],
)

fig.update_layout(
    title="Distribución de Ventas de Bebidas No Alcohólicas\n<br><i>por Presentación en 2024</i>",
    title_x=0.5,
    paper_bgcolor="rgba(0,0,0,0)",  # Transparent background
    plot_bgcolor="rgba(0,0,0,0)",
)

fig.write_image(
    "../img/00-macro/01-ventas-presentacion.png",
    scale=3,
)
fig
