==> explore/02-questions.ipynb <==
import pandas as pd
import plotly.graph_objects as go

df = pd.read_csv("../data/bebidas-no-alcoholicas-pivot.csv").query(
    "Mercado == 'Mexico'"
)

df

## ¿Cómo ha sido la evolución entre refrescos de colas y sabores?

cola = df.query("Sabor == 'COLA'").groupby("Fecha").sum().reset_index()
other = df.query("Sabor != 'COLA'").groupby("Fecha").sum().reset_index()

fig = go.Figure()

fig.add_trace(
    go.Scatter(
        name="Cola",
        mode="markers+lines",
        x=cola["Fecha"],
        y=cola["Vtas Valor"],
        marker_symbol="star",
    )
)

fig.add_trace(
    go.Scatter(
        name="Otros sabores",
        mode="markers+lines",
        x=other["Fecha"],
        y=other["Vtas Valor"],
        marker_symbol="star",
    )
)

title = "Ventas de Bebidas No Alcohólicas\n<br><i>2022 a 2024</i>"
fig.update_layout(
    title=title,
    title_x=0.5,
    yaxis_title="Ventas ($)",
    barcornerradius=5,
    paper_bgcolor="rgba(0,0,0,0)",  # Transparent background
    plot_bgcolor="rgba(0,0,0,0)",
)

fig.add_vrect(
    x0="2022-01-01",
    x1="2022-12-31",
    fillcolor="blue",
    opacity=0.1,
    layer="below",
    line_width=0,
)

fig.add_vrect(
    x0="2023-01-01",
    x1="2023-12-31",
    fillcolor="gray",
    opacity=0.1,
    layer="below",
    line_width=0,
)

fig.add_vrect(
    x0="2024-01-01",
    x1="2024-12-31",
    fillcolor="blue",
    opacity=0.1,
    layer="below",
    line_width=0,
)


fig.write_image(
    "../img/01-questions/01-ventas-mes.png",
    scale=3,
    width=1000,
    height=600,
)
fig.show()

## ¿Cuál es más sensible ante incrementos en precio?

cola = (
    df.query("Sabor == 'COLA'")
    .groupby("Fecha")
    .aggregate(
        {
            "Precio Unds Promedio": "mean",
            "Vtas Valor": "sum",
        }
    )
    .reset_index()
)

cola["Last Vtas Valor"] = cola["Vtas Valor"].shift(1)
cola["Last Precio Unds Promedio"] = cola["Precio Unds Promedio"].shift(1)

cola["Vtas Valor Growth"] = cola["Vtas Valor"] - cola["Last Vtas Valor"]
cola["Precio Unds Promedio Change"] = (
    cola["Precio Unds Promedio"] - cola["Last Precio Unds Promedio"]
).abs()

cola["Vtas Valor Growth per unit price change"] = (
    cola["Vtas Valor Growth"] / cola["Precio Unds Promedio Change"]
)

other = (
    df.query("Sabor != 'COLA'")
    .groupby("Fecha")
    .aggregate(
        {
            "Precio Unds Promedio": "mean",
            "Vtas Valor": "sum",
        }
    )
    .reset_index()
)

other["Last Vtas Valor"] = other["Vtas Valor"].shift(1)
other["Last Precio Unds Promedio"] = other["Precio Unds Promedio"].shift(1)

other["Vtas Valor Growth"] = other["Vtas Valor"] - other["Last Vtas Valor"]
other["Precio Unds Promedio Change"] = (
    other["Precio Unds Promedio"] - other["Last Precio Unds Promedio"]
).abs()

other["Vtas Valor Growth per unit price change"] = (
    other["Vtas Valor Growth"] / other["Precio Unds Promedio Change"]
)

cola

fig = go.Figure()

fig.add_trace(
    go.Scatter(
        name="Cola",
        mode="markers+lines",
        x=cola["Fecha"],
        y=cola["Vtas Valor Growth per unit price change"],
        marker_symbol="star",
    )
)

fig.add_trace(
    go.Scatter(
        name="Otros sabores",
        mode="markers+lines",
        x=other["Fecha"],
        y=other["Vtas Valor Growth per unit price change"],
        marker_symbol="star",
    )
)

title = "Crecimiento de Ventas por Cambio de Precios\n<br><i>2022 a 2024</i>"
fig.update_layout(
    title=title,
    title_x=0.5,
    yaxis_title="Ventas ($) / Cambio de precio",
    barcornerradius=5,
    paper_bgcolor="rgba(0,0,0,0)",  # Transparent background
    plot_bgcolor="rgba(0,0,0,0)",
)

fig.add_vrect(
    x0="2022-01-01",
    x1="2022-12-31",
    fillcolor="blue",
    opacity=0.1,
    layer="below",
    line_width=0,
)

fig.add_vrect(
    x0="2023-01-01",
    x1="2023-12-31",
    fillcolor="gray",
    opacity=0.1,
    layer="below",
    line_width=0,
)

fig.add_vrect(
    x0="2024-01-01",
    x1="2024-12-31",
    fillcolor="blue",
    opacity=0.1,
    layer="below",
    line_width=0,
)


fig.write_image(
    "../img/01-questions/02-sensibilidad-precios.png",
    scale=3,
    width=1000,
    height=600,
)

fig.show()

RELEVANT_MONTHS = [4, 5, 6, 7, 8]

cola = (
    df.query("Sabor == 'COLA'")
    .astype({"Fecha": "datetime64[ns]"})
    .query("Fecha.dt.month.isin(@RELEVANT_MONTHS)")
    .groupby("Fecha")
    .aggregate(
        {
            "Vtas Unds": "sum",
            "Vtas Valor": "sum",
        }
    )
    .reset_index()
)

cola


# cola = df.query("Sabor == 'COLA'").groupby("Fecha")["Vtas Valor"].sum().reset_index()
# cola["Last Vtas Valor"] = cola["Vtas Valor"].shift(1)
# cola["Diff Vtas Valor"] = cola["Vtas Valor"] - cola["Last Vtas Valor"]

# other = df.query("Sabor != 'COLA'").groupby("Fecha")["Vtas Valor"].sum().reset_index()
# other["Last Vtas Valor"] = other["Vtas Valor"].shift(1)
# other["Diff Vtas Valor"] = other["Vtas Valor"] - other["Last Vtas Valor"]

# fig = go.Figure()

# fig.add_trace(
#     go.Scatter(
#         name="Cola",
#         mode="markers+lines",
#         x=cola["Fecha"],
#         y=cola["Diff Vtas Valor"],
#         marker_symbol="star",
#     )
# )

# fig.add_trace(
#     go.Scatter(
#         name="Otros sabores",
#         mode="markers+lines",
#         x=other["Fecha"],
#         y=other["Diff Vtas Valor"],
#         marker_symbol="star",
#     )
# )

# title = "Crecimiento de Ventas por Cambio de Precios\n<br><i>2022 a 2024</i>"
# fig.update_layout(
#     title=title,
#     title_x=0.5,
#     yaxis_title="Ventas ($) / Cambio de precio",
#     barcornerradius=5,
#     paper_bgcolor="rgba(0,0,0,0)",  # Transparent background
#     plot_bgcolor="rgba(0,0,0,0)",
# )

# fig.add_vrect(
#     x0="2022-01-01",
#     x1="2022-12-31",
#     fillcolor="blue",
#     opacity=0.1,
#     layer="below",
#     line_width=0,
# )

# fig.add_vrect(
#     x0="2023-01-01",
#     x1="2023-12-31",
#     fillcolor="gray",
#     opacity=0.1,
#     layer="below",
#     line_width=0,
# )

# fig.add_vrect(
#     x0="2024-01-01",
#     x1="2024-12-31",
#     fillcolor="blue",
#     opacity=0.1,
#     layer="below",
#     line_width=0,
# )


# fig
